move overlay component
